refactor: consolidate cloud-init code into src/cloudinit
Move all cloud-init related files to a dedicated subdirectory to improve code
organization and make testing easier.
