+++ src/util.rs
@@ -1,7 +1,6 @@
 pub mod async_caller;
 pub mod hex;
 pub mod input;
-pub mod qemu;
 pub mod system_command;
 #[cfg(not(windows))]
 pub mod terminal;
@@ -9,7 +8,6 @@ pub mod terminal;
 pub use async_caller::*;
 pub use hex::*;
 pub use input::*;
-pub use qemu::*;
 pub use system_command::*;
 #[cfg(not(windows))]
 pub use terminal::*;
